+++ __tests__/fields.js
@@ -1,0 +1,34 @@
+'use strict';
+
+const Fields = require('../src/lib/fields');
+
+test('Fields Class', () => {
+  const pass = { structure: {} };
+  const fields = new Fields(pass, 'headerFields');
+  // should not add empty arrays if not needed
+  expect(Object.keys(pass.structure).length).toBe(0);
+  // add
+  fields.add('testKey', 'testLabel', 'testValue');
+  expect(fields.get('testKey')).toMatchObject({
+    key: 'testKey',
+    label: 'testLabel',
+    value: 'testValue',
+  });
+  // setValue
+  fields.setValue('testKey', 'newValue');
+  expect(fields.get('testKey')).toMatchObject({
+    key: 'testKey',
+    label: 'testLabel',
+    value: 'newValue',
+  });
+  // Add should replace the same key
+  fields.add('testKey', 'testLabel2', 'testValue2');
+  expect(fields.get('testKey')).toMatchObject({
+    key: 'testKey',
+    label: 'testLabel2',
+    value: 'testValue2',
+  });
+  // remove should remove the entry and whole key if it last one
+  fields.remove('testKey');
+  expect(Object.keys(pass.structure).length).toBe(0);
+});